mcp-quendoo-chatbot/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0


# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Encryption
cryptography==44.0.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.10

# Authentication
pyjwt==2.10.1
python-jose[cryptography]==3.3.0

# Environment
python-dotenv==1.0.1

# PostgreSQL (optional)
psycopg2-binary==2.9.10

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
